+++ R/as_fc.R
@@ -4,7 +4,7 @@
 #' @param .data Data frame to be initialised as a flowchart.
 #' @param N Number of rows of the study in case `.data` is NULL.
 #' @param label Character or expression with the text that will be shown in the box.
-#' @param text_pattern Structure that will have the text in the box. It recognizes label, n, N and perc within brackets. By default it is "\{label\}\\n\{N\}". If label is an expression, the label is always placed at the beginning of the pattern, followed by a line break where the structure specified by text_pattern is placed.
+#' @param text_pattern Structure that will have the text in the box. It recognizes label, n, N and perc within brackets. By default it is "\{label\}\\n\{N\}". If label is an expression or bold_n is TRUE, the label is always placed at the beginning of the pattern, followed by a line break where the structure specified by text_pattern is placed.
 #' @param just Justification for the text: left, center or right. Default is center.
 #' @param text_color Color of the text. It is black by default. See the `col` parameter for \code{\link{gpar}}.
 #' @param text_fs Font size of the text. It is 8 by default. See the `fontsize` parameter for \code{\link{gpar}}.
@@ -16,6 +16,7 @@
 #' @param width Width of the box. If NA, it automatically adjusts to the content (default). Must be an object of class \code{\link{unit}} or a number between 0 and 1.
 #' @param height Height of the box. If NA, it automatically adjusts to the content (default). Must be an object of class \code{\link{unit}} or a number between 0 and 1.
 #' @param hide Logical value to hide the initial box or not. Default is FALSE. hide = TRUE can only be combined with fc_split().
+#' @param bold_n Logical value to format the number in bold.
 #'
 #' @return List with the dataset and the initialized flowchart parameters.
 #'
@@ -26,7 +27,7 @@
 #'
 #' @export
 
-as_fc <- function(.data = NULL, N = NULL, label = "Initial dataframe", text_pattern = "{label}\n{N}", just = "center", text_color = "black", text_fs = 8, text_fface = 1, text_ffamily = NA, text_padding = 1, bg_fill = "white", border_color = "black", width = NA, height = NA, hide = FALSE) {
+as_fc <- function(.data = NULL, N = NULL, label = "Initial dataframe", text_pattern = "{label}\n{N}", just = "center", text_color = "black", text_fs = 8, text_fface = 1, text_ffamily = NA, text_padding = 1, bg_fill = "white", border_color = "black", width = NA, height = NA, hide = FALSE, bold_n = FALSE) {
 
   if(is.null(.data) & is.null(N)) {
     cli::cli_abort("Either {.arg .data} or {.arg N} arguments must be specified.")
@@ -64,22 +65,46 @@ as_fc <- function(.data = NULL, N = NULL, label = "Initial dataframe", text_patt
       bg_fill = bg_fill,
       border_color = border_color,
       width = width,
-      height = height
+      height = height,
+      end = TRUE
     )
 
     if(is.character(label)) {
 
-      new_fc <- new_fc |>
-        dplyr::mutate(text = as.character(stringr::str_glue(text_pattern)))
+      if(!bold_n) {
+
+        new_fc <- new_fc |>
+          dplyr::mutate(text = as.character(stringr::str_glue(text_pattern)))
+
+      } else {
+
+        text_pattern_exp <- gsub("\\{label\\}", "", text_pattern)
+
+        new_fc <- new_fc |>
+          dplyr::mutate(text = list(substitute(atop(x, bold(y)), list(x = label, y = stringr::str_glue(text_pattern_exp)))))
+
+      }
 
     } else {
 
       if(is.expression(label)) {
 
-        text_pattern_exp <- gsub("\\{label\\}", "", text_pattern)
+        if(!bold_n) {
 
-        new_fc <- new_fc |>
-          dplyr::mutate(text = list(substitute(atop(x, y), list(x = label[[1]], y = stringr::str_glue(text_pattern_exp)))))
+          text_pattern_exp <- gsub("\\{label\\}", "", text_pattern)
+
+          new_fc <- new_fc |>
+            dplyr::mutate(text = list(substitute(atop(x, y), list(x = label[[1]], y = stringr::str_glue(text_pattern_exp)))))
+
+        } else {
+
+          text_pattern_exp <- gsub("\\{label\\}", "", text_pattern)
+
+          new_fc <- new_fc |>
+            dplyr::mutate(text = list(substitute(atop(x, bold(y)), list(x = label[[1]], y = stringr::str_glue(text_pattern_exp)))))
+
+        }
+
 
       } else {